+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/RealMiniAppDisplaySpec.kt
@@ -9,10 +9,9 @@ import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.times
 import com.nhaarman.mockitokotlin2.verify
-import com.rakuten.tech.mobile.miniapp.MiniAppSdkException
+import com.rakuten.tech.mobile.miniapp.*
 import com.rakuten.tech.mobile.miniapp.TEST_HA_NAME
 import com.rakuten.tech.mobile.miniapp.TEST_MA
-import com.rakuten.tech.mobile.miniapp.TestActivity
 import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
@@ -43,7 +42,8 @@ class RealMiniAppDisplaySpec {
                 miniAppMessageBridge = miniAppMessageBridge,
                 miniAppNavigator = mock(),
                 hostAppUserAgentInfo = TEST_HA_NAME,
-                miniAppCustomPermissionCache = mock()
+                miniAppCustomPermissionCache = mock(),
+                queryParams = TEST_URL_PARAMS
             )
         }
     }
@@ -56,7 +56,8 @@ class RealMiniAppDisplaySpec {
             miniAppMessageBridge = miniAppMessageBridge,
             miniAppNavigator = mock(),
             hostAppUserAgentInfo = TEST_HA_NAME,
-            miniAppCustomPermissionCache = mock()
+            miniAppCustomPermissionCache = mock(),
+            queryParams = TEST_URL_PARAMS
         )
 
         realDisplay.miniAppInfo.apply {