+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayActivity.kt
@@ -125,7 +125,8 @@ class MiniAppDisplayActivity : BaseActivity() {
                 this@MiniAppDisplayActivity,
                 appUrl,
                 miniAppMessageBridge,
-                miniAppNavigator
+                miniAppNavigator,
+                AppSettings.instance.urlParameters
             )
         } else {
             viewModel.obtainMiniAppDisplay(
@@ -133,7 +134,8 @@ class MiniAppDisplayActivity : BaseActivity() {
                 appInfo,
                 appId!!,
                 miniAppMessageBridge,
-                miniAppNavigator
+                miniAppNavigator,
+                AppSettings.instance.urlParameters
             )
         }
     }